Template to treat 'secondary supplemental' as supplemental (#203)
Currently, `Template` treats "secondary supplemental" alignments as both
secondary alignments and as supplemental alignments. Before this PR,
`Template` would _**add**_ the "secondary supplemental" alignments into
both the secondary and supplementary lists, thereby duplicating them.
This behavior makes it difficult to use `Template` in complex alignment
scenarios since there are duplicated records across both the secondary
and supplementary lists e.g. records become duplicated when writing to a
BAM file with `template.write_to(writer)`.

It would be more intuitive to me that supplemental alignments of a
secondary alignment belong in the "supplementary" bin instead of both
the secondary and the supplementary bin. With this PR, the fields of
Template become more specific:

- Primary: all primary non-supplementary alignments
- Secondary: all secondary non-supplementary alignments
- Supplementary: all supplementary alignments

For context, I have a situation where I am intentionally creating and
comparing primary and secondary alignments, each having a chimeric
alignment (e.g. an alignment across different reference sequences), and
having to de-duplicate the supplemental records from both the secondary
and supplementary bins is error-prone and a lot of work.