Add a function for determining read pair orientation (#201)
I have a project where I am re-building a complex alignment situation
for a template and it involves an assessment of read pair orientation
and whether the new pair is "proper" or not.

The functions in this PR allow for an optional R2 such as when you are
iterating through a BAM in coordinate order and don't have easy access
to the R2. When this happens, R1 is required to have enough mate
information to derive the pair status and whether or not the read pair
is "proper".

The implementation honors what is implemented in HTSJDK/fgbio but with:

1. Fewer exceptions by using `None` instead of an exception, when
possible
2. Proper pair status also considers template length / insert size

---------

Co-authored-by: Tim Fennell <tfenne@tfenne.com>