Add methods to fix mate info on non-primaries and templates (#204)
Closes https://github.com/fulcrumgenomics/fgpyo/issues/83
Closes https://github.com/fulcrumgenomics/fgpyo/issues/182

I need a set of functions to make it easier to set mate info on
secondary, supplemental, "secondary supplemental", and templates. The
work in this PR builds on the work in:

- https://github.com/fulcrumgenomics/fgpyo/pull/201
- https://github.com/fulcrumgenomics/fgpyo/pull/202
- https://github.com/fulcrumgenomics/fgpyo/pull/203
- https://github.com/fulcrumgenomics/fgpyo/pull/205

This PR introduces:

-
[`set_mate_info_on_secondary`](https://github.com/fulcrumgenomics/fgpyo/blob/32024b9190aee61e02b96c89bc54c0ed6ac14cc1/fgpyo/sam/__init__.py#L879-L914)
-
[`set_mate_info_on_supplementary`](https://github.com/fulcrumgenomics/fgpyo/blob/32024b9190aee61e02b96c89bc54c0ed6ac14cc1/fgpyo/sam/__init__.py#L917C5-L948)
-
~[`set_mate_info_for_template`](https://github.com/fulcrumgenomics/fgpyo/blob/32024b9190aee61e02b96c89bc54c0ed6ac14cc1/fgpyo/sam/__init__.py#L1294-L1316)~
-
[`Template.set_mate_info()`](https://github.com/fulcrumgenomics/fgpyo/blob/32024b9190aee61e02b96c89bc54c0ed6ac14cc1/fgpyo/sam/__init__.py#L1232-L1234)

All "mate" fields are set using the primary of the opposite read
ordinal.

However, I am seeking advice on how we should set:

- `is_proper_pair`
- `template_length`

###### For Secondary Alignments

For secondary alignments, I personally am curious about those values not
as they are computed on the two primary alignments, but between the
secondary alignment and the primary of the opposite ordinal (e.g. if
this secondary _**was**_ primary, then what would it look like paired
with the primary of the opposite ordinal).

###### For Supplementary Alignments

For supplementary alignments of the _**primary**_ alignment, it makes
sense to have these two values point to the primary alignment of the
mate. However, for _**secondary**_ supplemental alignments, I think
these values are undefined without knowing more about the specific
alignment situation, so I am choosing to leave them as-is.