feat: `require()` (#246)
## Summary

Closes #132 

I've been using this implementation in a client repo and found it
helpful, so thought it would be good to transfer it over.

I do still think the utility presents a possible footgun for coverage
checks, since both the happy and unhappy paths share a single line of
code, and we may want to consider including a warning in the
documentation.